abis: bugfix. use latest genabi
It's possible for an event to have the same signature but different
number of indexed inputs; in which case the events should be treated as
distinct. Eg ERC20 and ERC721 have the same transfer event signature but
differ in that ERC20/Transfer's last input is not indexed and
ERC721/Transfer's last input is indexed.